Update README.md
The Epidemic routing algorithm was first introduced by Demers [for synchronization of replicated databases. Vahdat et al., later modified this protocol making it a flooding-base routing protocol for Delay Tolerant Networks (DTN). This routing protocol works on the theory of an Epidemic disease. That is, the message copy is flooded to all the encountered relay nodes in the network that do not have a copy of the same message on their buffer. In Epidemic routing protocol, each node maintains two buffers: one for storing the messages generated by the node itself and the other one for storing those messages received from the other nodes. Each message in the network is also assigned with a unique message identifier (message ID) and each node maintains a Summary Vector (SV) containing a list of all the message ID currently stored on its buffer. When two nodes come into a communication range, they exchange their SVs with each other. By comparing the SVs, the two nodes exchange all the messages that they do not have in common.

 To avoid multiple connections between the same pair of nodes, each node maintains a list of recently seen nodes and an SV is not exchanged with the node that has been seen within a predefined time period. The Epidemic protocol ensures a high delivery of messages by flooding message replicas in all the encountered nodes in the network; this ensures the optimal path for message delivery by exploring all the possible paths to the destination. As a result, this technique results in redundant messages against node failure. The major disadvantage of the Epidemic protocol is that it consumes a lot of network resources such as bandwidth, energy and buffer capacity when forwarding multiple copies of the same messages to relay nodes. As a result, there was a need todevelop more energy-efficient epidemic routing protocol.

 For the design of the proposed energy-efficient routing protocols, the same assumptions that were utilized in the original routing protocols are considered, but the following energy-efficient parameters are also introduced:

Minimum energy threshold (T): This is a value set by the network administrator (in Joules) to prevent relay nodes that have low energy level from being selected as next-hop nodes. Before a source node can forward a message copy to an encountered relay node, the energy level of the neighbor node is checked so that if it is less than the prescribed value T, the message copy will not be sent to that node. This mechanism assists in selecting only the nodes that have enough energy to deliver the message copy to the destination node or closer to the destination node. However, if the encountered relay node is the destination node, the message copy is forwarded directly to it regardless of the energy level of that node. 
One-hop acknowledgment mechanism: This is implemented to remove the copies of an already delivered message from the network and to prevent unnecessary transmission of such message. In this mechanism, when the message copy has been delivered to its destination, the destination node sends an acknowledgment message to the last encounter node that has sent the message. Both nodes update their acknowledgment tables and drop the message from their buffer to prevent it from further relaying. After a message has been dropped by the two nodes, both nodes update the acknowledgment table and flood their updated acknowledgment table to the network so that any node that comes in contact with a node having updated acknowledgment table can update their own acknowledgment table and delete the already delivered messages from their buffer.